requirements.txt
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
python-dotenv==1.2.1